*) Based on code from various people - Added: Catacombs classes spell handlers and support. *) Code from Uthgard - Bugfix: Fixed an issue in the MoveObject method which used the target table as the source table. - Bugfix: Fixed an issue where DuplicateNameCheckRequestHandler did not check archive accounts when checking if the account existed. *) Code from ZioRed - Bugfix: Resist buffs are no longer affected by spec, they will always be the value stated. - Bugfix: A few /item commands are no longer double escaped. *) Code from Me - Bugfix: Autoloot is by default set to true for new characters. - Added: Xfire support in game. Usage: /xfire on/off and relog, this will then display your character information in Xfire. - Changed: Savage spells no longer need new handlers for their spells, there is a new field in the Spell table for HealthPenalty. - Added: A field in the Style table for ArmorHitLocation, for special styles which specify where in the armor they will hit. (Vampiir) - Added: A new server property "allow_roam", set this to false if you don't want mobs to roam on the server. - Bugfix: Mobs will no longer try to roam if their speed is set to 0. - Removed: Restrictions for evade and parry requiring these skills before any check is done, some buffs increase the evade and parry chance (Vampiir). - Changed: Added a new property to GamePlayer, CanUseSlashLevel, this is set as the player is loaded, and is only assigned once, this replaces the check done on /level use and will check the account to see if it's eligible, this makes it possible to use the Mythic system for increasing XP for catacombs classes if they are eligible for /level. - Bugfix: Fixed an issue in RemoveAllSkills() method, previously it set the skill as null, instead of removing it, this caused errors with trades etc. - Added: A new server property for "allow_cata_slash_level", set this to true to allow them to /level, set to false to enable 50% xp boost up to level 20 if they are eligible. - Added: IsSilenced GameLiving property for the future Mauler spell handlers that are in development. - Bugfix: Attempted to fix an issue I've been encountering when I've been sending messages on a GM account, pressing reply after I've sent a message resulted in /send You appearing. - Added: Experimental support for Instances/ Task Dungeons that is still in development. - Removed: My ugly code for drop chances, and reverted to the old system by request of Dre. - Added: Dragon Kills now display in players stats. - Added: PropertyChangingAbility which can be used similar to realm ability equivalent, used for Vampiir's currently but has customisable power. - Added: SpellCastingAbilityHandler, an ability handler which casts spells, looking at this for the Vampiir Mana Leach bolt, but as of yet this does not work. - Added: When being promoted to a class which loses an ability, for example Heretic and losing leather, you will be checked in the equipment you have, and it will be removed if you cannot wear it. - Bugfix: /mob create command will no longer set the peace flag if the peace flag is already applied in the GameNPC's constructor, it will also not overwrite the other flags. - Bugfix: Various guild abilities, including the social windows show offline support will now work with archived characters.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@574 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServerScripts/gmcommands/debugjump.cs
@@ -66,7 +66,7 @@ namespace DOL.GS.Scripts
 				client.Out.SendMessage("Unknown region (" + RegionID.ToString() + ").", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 				return false;
 			}
-			else if (reg.Expansion >= client.ClientType)
+			else if (reg.Expansion >= (int)client.ClientType)
 			{
 				client.Out.SendMessage("Region (" + reg.Description + ") is not supported by your client.", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 				return false;
